Fix warnings when compiling for 64-bit ARM and iOS 7

+++ ALAlertBanner/ALAlertBanner.m
@@ -72,7 +72,7 @@ static CGFloat const kForceHideAnimationDuration = 0.1f;
 @implementation UIColor (LightAndDark)
 
 - (UIColor *)darkerColor {
-    float h, s, b, a;
+    CGFloat h, s, b, a;
     if ([self getHue:&h saturation:&s brightness:&b alpha:&a])
         return [UIColor colorWithHue:h saturation:s brightness:b * 0.75 alpha:a];
     return nil;
